+++ app/containers/Errors/constants.js
@@ -10,6 +10,8 @@
 
 export const HANDLE_SDK_ERROR = 'app/Errors/HANDLE_SDK_ERROR';
 export const SET_CRITICAL_ERROR = 'app/Errors/SET_CRITICAL_ERROR';
+export const SET_SESSION_ENDED_BY_SUPERVISOR =
+  'app/Errors/SET_SESSION_ENDED_BY_SUPERVISOR';
 export const SET_NON_CRITICAL_ERROR = 'app/Errors/SET_NON_CRITICAL_ERROR';
 export const DISMISS_ERROR = 'app/Errors/DISMISS_ERROR';
 export const ADD_STAT_ERROR_ID = 'app/Errors/ADD_STAT_ERROR';